http_api: Move hbit post code into index.rs
