+++ cnd/src/http_api/routes.rs
@@ -1,6 +1,5 @@
 pub mod index;
 pub mod peers;
-pub mod post;
 pub mod rfc003;
 
 use crate::{
